refactor: drag 코드 정리
