+++ packages/gdsc-dju-blog/src/pages/Home/onDrag.ts
@@ -1,0 +1,39 @@
+import React, { useState } from 'react';
+
+const onDrag = (scrollRef: React.RefObject<HTMLDivElement>) => {
+  const [isDrag, setIsDrag] = useState(false);
+  const [startX, setStartX] = useState(0);
+  const onDragStart = (e: React.MouseEvent<HTMLButtonElement>) => {
+    e.preventDefault();
+    setIsDrag(true);
+    if (scrollRef.current?.scrollLeft !== undefined)
+      setStartX(e.pageX + scrollRef.current.scrollLeft);
+  };
+  const onDragEnd = () => {
+    setIsDrag(false);
+  };
+  const onDragMove = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (isDrag) {
+      if (scrollRef.current !== null) {
+        const { scrollWidth, clientWidth, scrollLeft } = scrollRef.current;
+        if (scrollRef.current?.scrollLeft !== undefined)
+          scrollRef.current.scrollLeft = startX - e.pageX;
+
+        if (scrollLeft === 0) {
+          setStartX(e.pageX);
+        } else if (scrollWidth <= clientWidth + scrollLeft) {
+          setStartX(e.pageX + scrollLeft);
+        }
+      }
+    }
+  };
+
+  return {
+    isDrag,
+    onDragStart,
+    onDragEnd,
+    onDragMove,
+  };
+};
+
+export default onDrag;
